+++ src/CsrfGuard.php
@@ -20,9 +20,14 @@ use InvalidArgumentException;
 class CsrfGuard
 {
     /**
-     * @var array<mixed> Php session data from superglobal.
+     * @var string Token storage key name in session array.
      */
-    private array $session;
+    public const TOKEN_STORAGE = 'csrf-tokens';
+
+    /**
+     * @var array<mixed> Reference to php session data from superglobal.
+     */
+    private array $storage;
 
     /**
      * @var int Max number of tokens stored in session.
@@ -44,21 +49,24 @@ class CsrfGuard
      *
      * @throws RuntimeException  If instance is created without start session
      *                           before and if token strenght parameter is
-     *                           less than 16
+     *                           less than 16.
      */
     public function __construct(int $maxStorage, int $tokenStrength = 16)
     {
+        //check if session is started
         if (\session_status() === PHP_SESSION_NONE) {
             throw new RuntimeException('Session must be started before create instance.');
         }
 
+        //check for minimun token strenth
         if ($tokenStrength < 16) {
             throw new RuntimeException('The minimum CSRF token strength is 16.');
         }
 
-        $_SESSION['CSRF'] = $_SESSION['CSRF'] ?? [];
+        //initialize session storage
+        $_SESSION[self::TOKEN_STORAGE] = $_SESSION[self::TOKEN_STORAGE] ?? [];
 
-        $this->session = &$_SESSION;
+        $this->storage = &$_SESSION[self::TOKEN_STORAGE];
         $this->maxStorage = $maxStorage;
         $this->tokenStrength = $tokenStrength;
     }
@@ -70,11 +78,8 @@ class CsrfGuard
      */
     private function dequeue(array &$array): void
     {
-        $size = \count($array);
-
-        while ($size > $this->maxStorage) {
+        if (\count($array) > $this->maxStorage) {
             \array_shift($array);
-            $size--;
         }
     }
 
@@ -85,17 +90,18 @@ class CsrfGuard
      */
     public function getToken(): array
     {
+        //generate new token
         $token = $this->generateToken();
-
+        //pick the name of the token
         $name = $token['name'];
-
-        $this->session['CSRF'][$name] = $token;
+        //store the token
+        $this->storage[$name] = $token;
 
         //storage cleaning!
         //warning!! if you get in a page more token of maximun storage,
         //will there a leak of token, the firsts generated
         //in future I think throw and exception.
-        $this->dequeue($this->session['CSRF']);
+        $this->dequeue($this->storage);
 
         return $token;
     }
@@ -103,20 +109,24 @@ class CsrfGuard
     /**
      * Return timed csrf token as array.
      *
-     * @param int $ttl Time to live for the token.
+     * @param int $ttl Time to live for the token, default 600 -> 10 minutes.
      *
      * @return array<mixed>
      */
-    public function getTimedToken(int $ttl): array
+    public function getTimedToken(int $ttl = 600): array
     {
+        //generate new token
         $token = $this->generateToken();
+        //store token expiration
+        //add new key to token array
         $token['time'] = \time() + $ttl;
-
+        //pick the name of the token
         $name = $token['name'];
+        //store the token
+        $this->storage[$name] = $token;
 
-        $this->session['CSRF'][$name] = $token;
-
-        $this->dequeue($this->session['CSRF']);
+        //storage cleaning
+        $this->dequeue($this->storage);
 
         return $token;
     }
@@ -128,7 +138,9 @@ class CsrfGuard
      */
     private function generateToken(): array
     {
+        //generate a random token name
         $name = 'csrf_'.\bin2hex(\random_bytes(8));
+        //generate a random token value
         $value = \bin2hex(\random_bytes($this->tokenStrength));
 
         return ['name' => $name, 'value' => $value];
@@ -161,24 +173,21 @@ class CsrfGuard
      */
     private function doChecks(string $value, string $key): bool
     {
-        $tokens = &$this->session['CSRF'];
-
-        return $this->tokenIsValid($tokens, $value, $key) &&
-               $this->tokenIsExiperd($tokens, $key)  &&
-               $this->deleteToken($tokens, $key);
+        return $this->tokenIsValid($value, $key) &&
+               $this->tokenIsExpired($key)  &&
+               $this->deleteToken($key);
     }
 
     /**
      * Delete token after validation.
      *
-     * @param array<mixed>  $tokens
-     * @param string        $key
+     * @param string $key
      *
      * @return bool
      */
-    private function deleteToken(array &$tokens, string &$key): bool
+    private function deleteToken(string &$key): bool
     {
-        unset($tokens[$key]);
+        unset($this->storage[$key]);
 
         return true;
     }
@@ -186,34 +195,32 @@ class CsrfGuard
     /**
      * Check if token is valid
      *
-     * @param array<mixed>  $tokens
-     * @param string        $value
-     * @param string        $key
+     * @param string $value
+     * @param string $key
      *
      * @return bool
      */
-    private function tokenIsValid(array &$tokens, string &$value, string &$key): bool
+    private function tokenIsValid(string &$value, string &$key): bool
     {
-        //if token is not existed
-        if (empty($tokens[$key])) {
+        //if token doesn't exist
+        if (empty($this->storage[$key])) {
             return false;
         }
 
-        return \hash_equals($tokens[$key]['value'], $value);
+        return \hash_equals($this->storage[$key]['value'], $value);
     }
 
     /**
      * Check if timed token is expired.
      *
-     * @param array<mixed>  $tokens
-     * @param string        $key
+     * @param string $key
      *
      * @return bool
      */
-    private function tokenIsExiperd(array &$tokens, string &$key): bool
+    private function tokenIsExpired(string &$key): bool
     {
         //if timed and if time is valid
-        if (isset($tokens[$key]['time']) && $tokens[$key]['time'] < \time()) {
+        if (isset($this->storage[$key]['time']) && $this->storage[$key]['time'] < \time()) {
             return false;
         }
 
@@ -227,7 +234,7 @@ class CsrfGuard
      */
     public function garbageCollector(int $preserve): void
     {
-        if ($this->maxStorage === \count($this->session['CSRF'])) {
+        if ($this->maxStorage === \count($this->storage)) {
             $this->cleanStorage($preserve);
         }
     }
@@ -249,7 +256,7 @@ class CsrfGuard
      *
      * @throws InvalidArgumentException If arguments lesser than 0 or grater than max storage value.
      */
-    private function cleanStorage(int $preserve): void
+    private function cleanStorage(int $preserve = 0): void
     {
         if ($preserve < 0) {
             throw new InvalidArgumentException('Argument value should be grater than zero.');
@@ -259,7 +266,6 @@ class CsrfGuard
             throw new InvalidArgumentException("Argument value should be lesser than max storage value ({$this->maxStorage}).");
         }
 
-        $tokens = &$this->session['CSRF'];
-        $tokens = \array_splice($tokens, -$preserve);
+        $this->storage = \array_splice($this->storage, -$preserve);
     }
 }