+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [v1.4.0](https://github.com/linna/csrf-guard/compare/v1.3.2...v1.4.0) - 2020-XX-XX
 
-### Changed
+### Added
 * PHP 8.0 support
+
+### Changed
+* `getTimedToken()` now `$ttl` parameter has a default value of 600 (seconds)
+* Minor code optimization
 * Minimun PHP version: 7.4
 
 ## [v1.3.2](https://github.com/linna/csrf-guard/compare/v1.3.1...v1.3.2) - 2020-02-11